Refactored components to use new img file
