+++ projects/frontend/src/components/index.js
@@ -5,78 +5,12 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import React from 'react';
 import PropTypes from 'prop-types';
-import Img from 'gatsby-image';
-
 import styled, { css, keyframes } from 'styled-components';
 
+import Img from 'gatsby-image';
+
 import { Center, Rounded, ZDepth1, ZDepth3 } from './mixins';
-
-export const imgPropType = PropTypes.shape({
-  aspectRatio: PropTypes.number,
-  base64: PropTypes.string,
-  sizes: PropTypes.string,
-  src: PropTypes.string,
-  srcSet: PropTypes.string,
-  srcSetWebp: PropTypes.string,
-  srcWebp: PropTypes.string,
-});
-
-export const imgPropTypeShape = PropTypes.shape({ sizes: imgPropType });
-
-export const imgListPropType = PropTypes.shape({
-  edges: PropTypes.arrayOf(
-    PropTypes.shape({
-      node: PropTypes.shape({
-        id: PropTypes.string,
-        sizes: imgPropType,
-      }),
-    })
-  ),
-});
-
-const BgWrap = styled.div`
-  position: relative;
-  // Make sure children display on top of the image
-  & > * {
-    z-index: 1;
-  }
-`;
-
-export const BackgroundImg = ({ img, children, ...props }) => (
-  <BgWrap {...props}>
-    {img && (
-      <Img
-        style={{
-          position: 'absolute',
-          left: 0,
-          top: 0,
-          zIndex: 0,
-          width: '100%',
-          height: '100%',
-        }}
-        {...img}
-      />
-    )}
-    {children}
-  </BgWrap>
-);
-
-BackgroundImg.propTypes = {
-  img: PropTypes.shape({
-    resolutions: PropTypes.object,
-    sizes: PropTypes.object,
-  }),
-  // eslint-disable-next-line react/forbid-prop-types
-  props: PropTypes.object,
-  children: PropTypes.node,
-};
-
-export const Avatar = styled(Img)`
-  border-radius: 100%;
-  margin-right: 0.5em;
-  height: 5vh;
-  width: 5vh;
-`;
+import { BackgroundImg } from './Img';
 
 export const Banner = styled.section`
   display: flex;
@@ -134,6 +68,20 @@ const FormField = styled.div`
   }
 `;
 
+const Input = ({ title, value = '', ...props }) => (
+  <FormField>
+    {title && <p>{title}</p>}
+    <input {...props} value={value} />
+  </FormField>
+);
+
+Input.propTypes = {
+  title: PropTypes.string,
+  value: PropTypes.string,
+};
+
+export { Input };
+
 const IconButton = ({ img, ...props }) => (
   <IconButtonStyle {...props}>
     <Icon {...img} />
@@ -148,7 +96,8 @@ IconButton.propTypes = {
 };
 
 const IconButtonStyle = styled.button`
-  ${Center} background-color: transparent;
+  ${Center};
+  background-color: transparent;
   height: 5vh;
   width: 5vh;
   border: none;
@@ -161,20 +110,6 @@ const Icon = styled(Img)`
 `;
 
 export { IconButton };
-
-const Input = ({ title, value = '', ...props }) => (
-  <FormField>
-    {title && <p>{title}</p>}
-    <input {...props} value={value} />
-  </FormField>
-);
-
-Input.propTypes = {
-  title: PropTypes.string,
-  value: PropTypes.string,
-};
-
-export { Input };
 
 const TextArea = ({ title, ...props }) => (
   <FormField>
@@ -232,12 +167,3 @@ export const SplitSection = styled(BackgroundImg)`
     }
   }
 `;
-
-export const imgMatch = (imgs, match) =>
-  imgs.edges.find(({ node: { id } }) => id.includes(match)).node;
-
-export const mergeSocial = (socials, imgs) =>
-  socials.map(social => ({
-    ...social,
-    img: imgMatch(imgs, social.name),
-  }));
